pyproject.toml
[project]
name = "counselai"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    # Web app
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "gradio>=5.47.2,<6.0",
    # LLM
    "openai>=1.109.1",
    # Auth (Clerk JWT verification)
    "pyjwt[crypto]>=2.10",
    "httpx>=0.28",
    # Storage
    "boto3>=1.40",
    # Config
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "pytest>=8.3",
    "ipykernel>=6.30.1",  # so the reference notebook still opens
]

[tool.uv]
package = false
